pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-governance-bootstrap"
version = "1.0.0"
description = "AI Governance Bootstrap - Governed worker execution framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.scripts]
run-bootstrap = "runtime.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["runtime*", "bootstrap*", "agents*", "constitution*", "guardrails*", "learning*"]
